added script in sales to update ob when entry made for prev fyear
