+++ routes/common.js
@@ -1,6 +1,8 @@
 const rate = require('../models/rate');
 const sales = require('../models/sales');
 const discountRoute = require('./discount');
+const mongoose = require('mongoose');
+const {ObjectId} = require('mongodb');
 
 let common = {};
 common.firstmethod = function(){
@@ -126,5 +128,159 @@ common.getRateList = function(){
             }
         });
     });
+}
+
+common.fetchCreditDebitsByYear = async function(year,user) {
+    // console.log("called!!");
+    year = parseInt(year);
+    const sales = mongoose.model("Sales");
+    const credits = await sales.aggregate([
+        {$match:{
+            $expr:{
+                $and:[
+                    {$eq: ["$customer_id", user._id]},
+                    {$eq: ["$is_active", "YES"]},
+                    {$eq: ["$is_delete", "NO"]},
+                    {$eq: ["$payment_type", "CREDIT"]},
+                    {$eq: ["$financial_year", year]},
+                    // {$gte: ["$sale_date", startDate]},
+                    // {$lte: ["$sale_date", endDate]},
+                ]
+            }
+        }},
+        {$group:{
+            _id: '$customer_id',
+            t_credit: {
+                $sum: '$total_amount'
+            }
+        }}
+    ]).exec(); 
+
+    const payments = mongoose.model("Payment");
+    const debits = await payments.aggregate([
+        {$match:{
+            $expr:{
+                $and:[
+                    {$eq: ["$customer_id", user._id]},
+                    {$eq: ["$is_active", "YES"]},
+                    {$eq: ["$is_delete", "NO"]},
+                    {$eq: ["$financial_year", year]},
+                    // {$gte: ["$createdAt", startDate]},
+                    // {$lte: ["$createdAt", endDate]},
+                ]
+            }
+        }},
+        {$group:{
+            _id: '$customer_id',
+            t_debit: {
+                $sum: '$amount'
+            }
+        }}
+    ]).exec();
+    const ob = mongoose.model("OpeningBalance");
+    const obs = await ob.aggregate([
+        {$match:{
+            $expr:{
+                $and:[
+                    {$eq: ["$customer_id", user._id]},
+                    {$eq: ["$is_active", "YES"]},
+                    {$eq: ["$is_delete", "NO"]},
+                    {$eq: ["$financial_year", year]},
+                    // {$gte: ["$createdAt", startDate]},
+                    // {$lte: ["$createdAt", endDate]},
+                ]
+            }
+        }},
+        {$group:{
+            _id: '$customer_id',
+            t_ob: {
+                $sum: '$amount'
+            }
+        }}
+    ]).exec();
+    return Promise.all([credits,debits,obs]); 
+}
+
+common.generateUserObForEachYear = async function(user, year){
+    let _resp = {
+        code : 201,
+        message : "Error Occurred",
+        data: []
+    };
+    let customerObj = {};
+    if(!!user)
+    {
+        customerObj = { _id: ObjectId(user) }
+    }
+    
+    return new Promise((resolve, reject) => {
+        const customer = mongoose.model("Customer");
+        customer.find(customerObj, function(err, users) {
+            if(err){
+                _resp.message = "Error in customer";
+                reject(_resp);
+            }else{
+                let fYear = year;
+                let temp = [];
+                let promiseArr = users.map((user)=>{
+                    return common.fetchCreditDebitsByYear(fYear, user).then((result) => {
+                        return result;
+                    });
+                })
+                Promise.all(promiseArr).then(function(results){
+                    // console.log(results);
+                    results.forEach((result) => {
+                        if (result.length > 0) {
+                            let outstanding = 0;
+                            const [credits, debits, obs] = result;
+                            // console.log(credits,debits,obs);
+                            const credit = credits.length > 0 ? credits[0].t_credit : 0;
+                            const debit = debits.length > 0 ? debits[0].t_debit : 0;
+                            const ob = obs.length > 0 ? obs[0].t_ob : 0;
+                            const nextFYear = parseInt(fYear)+1;
+                            outstanding = (parseInt(credit) + parseInt(ob)) - parseInt(debit);
+                            
+                            let cid;
+                            if( credits.length > 0 && credits[0]._id){
+                                cid = credits[0]._id;
+                            }else if(debits.length > 0 && debits[0]._id){
+                                cid = debits[0]._id;
+                            }else if(obs.length > 0 && obs[0]._id){
+                                cid = obs[0]._id;
+                            }
+    
+                            // console.log("outstanding::",cid,outstanding);
+                            if (outstanding >= 1) {
+                                temp.push({
+                                    "updateOne": {
+                                        "filter": { "customer_id": cid, "financial_year": nextFYear },
+                                        "update": { "$set": { "financial_year": nextFYear, "customer_id": cid, "amount": outstanding } },
+                                        "upsert": true,
+                                        "setDefaultsOnInsert": true
+                                    }
+                                });
+                            }   
+                        }
+                    });
+                    if(temp.length > 0){
+                        // res.json(temp);
+                        const ob = mongoose.model("OpeningBalance");
+                        ob.bulkWrite(temp,  { setDefaultsOnInsert: true }).then(state => {
+                            _resp.code = 200;
+                            _resp.message = "success";
+                            resolve(_resp);
+                        });
+                    }else{
+                        _resp.message = "no records found";
+                        reject(_resp);
+                    } 
+                }).catch(function(err){
+                    _resp.message = "Error in promise all";
+                    reject(_resp);
+                });
+            }
+        });
+    }); 
+    
 }
 module.exports = common;
